refactor(feature): extract file readers (read_file/read_parquet/iter_features/open_arrow) to _read (ARC-36)
Move the FeatureCollection file-reader method bodies and their private helpers
into pyramids.feature._read as free functions, leaving thin classmethod facades
that keep the full docstrings and doctests. Extracted:

- read_file / read_parquet (eager + dask backends) and the shared backend
  dispatch helpers _resolve_lazy_partitioning, _require_pyarrow, _compact,
  _import_dask_geopandas
- iter_features (streaming) with build_iter_read_kwargs / emit_features
- open_arrow (pyarrow RecordBatchReader)
- the _read_parquet_dask backend

The tunable _LAZY_TARGET_BYTES_PER_PARTITION knob moves with
_resolve_lazy_partitioning into _read; configure_lazy_vector (base/_configure.py)
now patches it there, and the partitioning / configure-lazy-vector tests and the
_require_pyarrow mocks are repointed to _read. _VALID_TILE_STRATEGIES stays a
class attribute (part of the public contract).

Second of the reader-extraction steps for the FeatureCollection god-class
decomposition (ARC-36 Option B).
